Allow downloading multiple files from Canvas
Previously the marker only allowed downloading a single attachment
and renamed it to `file_name` according to the configuration. Now,
leaving out the `file_name` option will download all the files
from the Canvas submission and retain the original names that the
files were submitted with.

If this option is used the user needs to be careful about ensuring
themselves that the possibly different file names are handled
properly, since Canvas does not let you restrict the names of the
submitted files like MarkUs.`