Update `marker.h` C template
The `marker.h` file now includes the ability to run a test in a
separate process to avoid letting it exit early with `exit(0)` to
fake-pass the test.

The parent (main) test process expects the child to write some value
into the pipe when it is done running, and only counts the test as
passed if it sees this exact value.