Can now query LMS for submission URL
This allows us to ask the LMS object for a URL that would open up the submission page for a student. Can be useful for quickly opening up a submission to check if everything was uploaded correctly.

This only works for Canvas for now, since MarkUs currently doesn't allow us to get the actual `result_id` which is needed to access the submission page.

Lastly, the URL is only returned if the LMS has already fetched the students and created the mapping. This is because the marker may be used without an LMS, so attempting to create a mapping manually would cause errors we don't want in those cases.